Add enums `stype` and `ltype` (#472)
- Expose enums at datatable's top level
- Add documentation for the enums
- Add exhaustive unit tests for those enums
- Remove stype ST_DATETIME_I8_PRTMN (the "epoch" datetime is just as good)
- Fix 2 warnings in column_fw.cc and column_from_pylist.cc

+++ c/column_fw.cc
@@ -36,7 +36,7 @@ void FwColumn<T>::replace_buffer(MemoryBuffer* new_mbuf, MemoryBuffer*) {
   MemoryBuffer* t = new_mbuf->shallowcopy();
   if (mbuf) mbuf->release();
   mbuf = t;
-  nrows = mbuf->size() / sizeof(T);
+  nrows = static_cast<int64_t>(mbuf->size() / sizeof(T));
 }
